Add a utilities file for the autoformatter
